Move ctutil from integration to core
